feat: rename ResponsesProvider to StorageResponsesProvider
